리눅스 서버와 유니티 게임 틀 완성
이제 게임 연출이나 UI 좀 다듬으면 될 것 같습니다

+++ server/client.c
@@ -1,32 +1,61 @@
-#include<stdio.h>
-#include<stdlib.h>
-#include<unistd.h>
-#include<signal.h>
-#include<fcntl.h>
-#include<sys/stat.h>
-#include<sys/ipc.h>
-#include<sys/shm.h>
-#include<string.h>
-#include "shared_data.h"
-
-#define FIFO_PATH "./client_ready_fifo"
+#include "client.h"
 
 SharedData* shmaddr;
 int clientID;
+int clientSocket;
 
-void signal_handler(int sig) {
-	int choice;
-	printf("선택할 숫자를 입력: ");
-	scanf("%d", &choice);
+void next_turn(int sig) {
 
-	shmaddr->choice[clientID] = choice;
+	//유니티로 현재 폭탄 든 사람 번호랑 남아있는 타이머카운트  전달.
+	// (shmaddr->bomb_owner - clientID + CLIENTS_NUM) % CLIENTS_NUM 번호로 전달. 유니티에서는 client 본인이 0번이 되게하려고
+	// shmaddr->timer 전달
 
-	kill(getppid(), SIGUSR1);
+	char buffer[sizeof(int) + sizeof(float)];
+
+	int bomb_owner = (shmaddr->bomb_owner - clientID + CLIENTS_NUM) % CLIENTS_NUM;
+
+	memcpy(buffer, &bomb_owner, sizeof(int));
+	memcpy(buffer + sizeof(int), &shmaddr->timer, sizeof(float));
+
+	write(clientSocket, buffer, sizeof(buffer));
+	printf("%d 플레이어 유니티로 데이터 전송 성공\n", clientID);
+
+	if(clientID == shmaddr->bomb_owner) {  //본인의 차례였을 경우	
+			
+		//유니티에게서 남은 타이머 카운트 받기 전까지 계속 대기
+		//카운트 받으면 공유메모리 갱신하고 메인으로 시그널 보냄
+
+		char buf[sizeof(float)];
+		memset(buf, 0, sizeof(buf));
+		int bytes_received = read(clientSocket, buf, sizeof(float));
+		if(bytes_received > 0) {
+			memcpy(&shmaddr->timer, buf, sizeof(float));
+			printf("클라이언트 %d의 턴 종료\n", clientID);
+		}
+
+		kill(getppid(), SIGUSR1);
+	} 
+	
 }
 
-int main(int argc, char *argv[]) {
+void game_end(int sig) {
 
-	clientID = atoi(argv[1]);
+	char buffer[sizeof(int) + sizeof(float)];
+
+	int bomb_owner = -1;
+	float timer = -1;
+
+	memcpy(buffer, &bomb_owner, sizeof(int));
+	memcpy(buffer + sizeof(int), &timer, sizeof(float));
+
+	write(clientSocket, buffer, sizeof(buffer));
+
+}
+
+void handle_client(int client_id, int client_socket) {
+
+	clientSocket = client_socket;
+	clientID = client_id;
 	printf("클라이언트 프로세스 %d  실행\n", clientID);
 	
 	key_t key;
@@ -35,22 +64,20 @@ int main(int argc, char *argv[]) {
 	int shmid = shmget(key, SHM_SIZE, IPC_CREAT | 0644);
 	shmaddr = (SharedData *)shmat(shmid, NULL, 0);	
 
-	int write_fd = open(FIFO_PATH, O_WRONLY);
-    if (write_fd == -1) {
+	int write_fd;
+	while((write_fd = open(FIFO_PATH, O_WRONLY)) == -1) {
         perror("write_fd 열기 실패");
-        return 1;
-    } 
-	/*
-	else {
-		printf("클라이언트 %d write_fd : %d", clientID, write_fd);
+		printf("%d write_fd 다시 시도 중\n", clientID);
+        sleep(1); 
 	}
-	*/
 
-	signal(SIGUSR1, signal_handler);
+	signal(SIGUSR1, next_turn);
+	signal(SIGUSR2, game_end);
 
 	const char *msg = "준비완료";
 
 	write(write_fd, msg, strlen(msg)); //열어뒀던 fifo에 메시지 쓰기	
+	printf("%d 준비완료 메시지 전송\n", clientID);
 
 	close(write_fd);
 
@@ -61,5 +88,4 @@ int main(int argc, char *argv[]) {
 
 	shmdt(shmaddr);
 
-	return 0;
 }
